refactor: acceptance 패키지에서 모든 인수테스트 관리
