+++ backend/src/test/java/com/woowacourse/kkogkkog/acceptance/ReservationAcceptanceTest.java
@@ -1,18 +1,18 @@
-package com.woowacourse.kkogkkog.reservation.acceptance;
+package com.woowacourse.kkogkkog.acceptance;
 
-import static com.woowacourse.kkogkkog.acceptance.AcceptanceContext.invokePostWithToken;
-import static com.woowacourse.kkogkkog.acceptance.AcceptanceContext.invokePutWithToken;
+import static com.woowacourse.kkogkkog.acceptance.support.AcceptanceContext.invokePostWithToken;
+import static com.woowacourse.kkogkkog.acceptance.support.AcceptanceContext.invokePutWithToken;
 import static com.woowacourse.kkogkkog.acceptance.AuthAcceptanceTest.회원가입을_하고;
+import static com.woowacourse.kkogkkog.acceptance.CouponAcceptanceTest.쿠폰_생성을_요청하고;
 import static com.woowacourse.kkogkkog.support.fixture.domain.MemberFixture.LEO;
 import static com.woowacourse.kkogkkog.support.fixture.domain.MemberFixture.ROOKIE;
 import static com.woowacourse.kkogkkog.support.fixture.dto.CouponDtoFixture.COFFEE_쿠폰_생성_요청;
 import static com.woowacourse.kkogkkog.support.fixture.dto.ReservationDtoFixture.예약_변경_요청;
 import static com.woowacourse.kkogkkog.support.fixture.dto.ReservationDtoFixture.예약_생성_요청;
-import static com.woowacourse.kkogkkog.coupon.acceptance.CouponAcceptanceTest.쿠폰_생성을_요청하고;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-import com.woowacourse.kkogkkog.acceptance.AcceptanceTest;
+import com.woowacourse.kkogkkog.acceptance.support.AcceptanceTest;
 import com.woowacourse.kkogkkog.reservation.presentation.dto.ReservationCreateRequest;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
@@ -21,6 +21,7 @@ import java.util.List;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 
+@SuppressWarnings("NonAsciiCharacters")
 public class ReservationAcceptanceTest extends AcceptanceTest {
 
     @Test
@@ -53,15 +54,15 @@ public class ReservationAcceptanceTest extends AcceptanceTest {
         );
     }
 
-    public ExtractableResponse<Response> 예약을_변경한다(Long id, String token, Object data) {
+    static ExtractableResponse<Response> 예약을_변경한다(Long id, String token, Object data) {
         return invokePutWithToken("/api/reservations/" + id, token, data);
     }
 
-    public static void 예약을_신청하고(String token, ReservationCreateRequest request) {
+    static void 예약을_신청하고(String token, ReservationCreateRequest request) {
         invokePostWithToken("/api/reservations", token, request);
     }
 
-    public static ExtractableResponse<Response> 예약을_신청한다(String token, Object data) {
+    static ExtractableResponse<Response> 예약을_신청한다(String token, Object data) {
         return invokePostWithToken("/api/reservations", token, data);
     }
 }
